Revert "Ovl: Add En_Test4"
This reverts commit 4019d5770d13725d655738d6feddeda33da36fdd.
