+++ packages/core/CMakeLists.txt
@@ -143,7 +143,6 @@ add_actor(oot En_Mag      "oot/ovl/actors/En_Mag/En_Mag.c")
 add_actor(mm Obj_Tsubo          "mm/ovl/actors/Obj_Tsubo/Obj_Tsubo.c")
 add_actor(mm En_Hs              "mm/ovl/actors/En_Hs/En_Hs.c")
 add_actor(mm En_Nwc             "mm/ovl/actors/En_Nwc/En_Nwc.c")
-add_actor(mm En_Test4           "mm/ovl/actors/En_Test4/En_Test4.c")
 add_actor(mm Custom_Magic_Fire  "mm/ovl/actors/Custom_Magic_Fire/Custom_Magic_Fire.c")
 add_actor(mm Custom_Magic_Wind  "mm/ovl/actors/Custom_Magic_Wind/Custom_Magic_Wind.c")
 add_actor(mm Custom_Magic_Dark  "mm/ovl/actors/Custom_Magic_Dark/Custom_Magic_Dark.c")
