+++ packages/core/include/combo/interface.h
@@ -7,7 +7,6 @@ void Interface_Update(GameState_Play* play);
 
 #if defined(GAME_MM)
 void Interface_NewDay(GameState_Play* play, int day);
-void Interface_SetHudVisibility(u16 hudVisibility);
 #endif
 
 #endif
