+++ packages/core/src/link_ovl.ld
@@ -1,6 +1,6 @@
 SECTIONS
 {
-  . = 0x80d00000;
+  . = 0x80800000;
 
   .text :   ALIGN(16) { SORT(*)(.text*); }
   .data :   ALIGN(16) { SORT(*)(.data*); SORT(*)(.bss*); }
